// resources/js/admin/guide-quests/api/definitions/guide-quest-detail-definition.ts
import { GuideQuestContentBlockDefinition } from './guide-quest-definition';

export interface GuideQuestBatchCraftedItemRequirementDefinition {
  requirement_index: number;
  source: string;
  item_id: number;
  name: string;
  type: string;
  type_name: string;
  amount: number;
  must_be_enchanted: boolean;
}

export default interface GuideQuestDetailDefinition {
  id: number;
  name: string;
  intro_text: GuideQuestContentBlockDefinition[] | null;
  desktop_instructions: GuideQuestContentBlockDefinition[] | null;
  mobile_instructions: GuideQuestContentBlockDefinition[] | null;
  required_level: number | null;
  required_reincarnation_amount: number | null;
  unlock_at_level: number | null;
  required_skill: number | null;
  skill_name: string | null;
  required_skill_level: number | null;
  required_secondary_skill: number | null;
  secondary_skill_name: string | null;
  required_secondary_skill_level: number | null;
  required_skill_type: number | null;
  skill_type_name: string | null;
  required_skill_type_level: number | null;
  required_passive_skill: number | null;
  passive_name: string | null;
  required_passive_level: number | null;
  required_class_specials_equipped: number | null;
  required_class_rank_level: number | null;
  required_specialty_type: string | null;
  required_fame_level: number | null;
  required_faction_id: number | null;
  faction_name: string | null;
  required_faction_level: number | null;
  must_be_pledged_to_faction: boolean | null;
  must_be_assisting_npc: boolean | null;
  required_game_map_id: number | null;
  game_map_name: string | null;
  be_on_game_map: number | null;
  required_to_be_on_game_map_name: string | null;
  required_quest_id: number | null;
  quest_name: string | null;
  parent_id: number | null;
  parent_quest_name: string | null;
  required_quest_item_id: number | null;
  quest_item_name: string | null;
  secondary_quest_item_id: number | null;
  secondary_quest_item_name: string | null;
  only_during_event: number | null;
  required_event_goal_participation: number | null;
  required_event_goal_crafting_participation: number | null;
  required_event_goal_enchanting_participation: number | null;
  required_kingdoms: number | null;
  required_kingdom_level: number | null;
  required_kingdom_units: number | null;
  required_kingdom_building_id: number | null;
  kingdom_building_name: string | null;
  required_kingdom_building_level: number | null;
  required_gold_bars: number | null;
  required_stats: number | null;
  required_str: number | null;
  required_dex: number | null;
  required_int: number | null;
  required_dur: number | null;
  required_chr: number | null;
  required_agi: number | null;
  required_focus: number | null;
  required_gold: number | null;
  required_gold_dust: number | null;
  required_shards: number | null;
  required_copper_coins: number | null;
  required_holy_stacks: number | null;
  required_attached_gems: number | null;
  required_delve_survival_time: number | null;
  required_delve_pack_size: number | null;
  required_batch_crafting_type: string | null;
  required_batch_crafting_type_name: string | null;
  required_batch_crafting_hours: number | null;
  required_batch_crafted_item_names: GuideQuestBatchCraftedItemRequirementDefinition[];
  xp_reward: number | null;
  gold_reward: number | null;
  gold_dust_reward: number | null;
  shards_reward: number | null;
  faction_points_per_kill: number | null;
}